refactor(tee): split tee/mod.rs god object into responsibility modules
tee/mod.rs (1483 LOC) -> extracted behavior modules, keeping the shared TEE
types (TeeType, TeeConfig, AttestationReport, TeeDeployParams, nonce helpers)
and the existing sibling mod decls in mod.rs (239 LOC): backend (TeeBackend
trait + registry), sidecar_attest (fetch/validate sidecar attestation),
verification (verdict types + quote-signature checks + measurements),
verify_flow (end-to-end verify + sidecar health/key/secret ops). The inline
pub mod mock and the unit tests move to tee/mock.rs and tee/tests.rs. Module
names avoid colliding with existing siblings (attestation, verify,
backend_factory). External API preserved via glob re-export at original
visibility (pub items pub, pub(crate) items pub(crate)); mock stays pub under
its cfg(test|test-utils) gate. Pure code motion, no behavior change; max
extracted module 175 LOC.
