+++ scripts/.god-object-baseline.txt
@@ -2,7 +2,6 @@
 # Regenerate after a real split: scripts/check-file-sizes.sh --update
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
 2966 ui/src/routes/create.tsx
-1483 sandbox-runtime/src/tee/mod.rs
 1281 ui/src/routes/sandboxes.$id.tsx
 1111 ai-agent-sandbox-blueprint-bin/src/main.rs
 1106 ui/src/routes/instances.$id.tsx
